+++ module/actor/sheets/base.js
@@ -81,6 +81,7 @@ export class ActorSheetPF extends ActorSheet {
       i.data.hasDamage = i.hasDamage;
       i.data.hasEffect = i.hasEffect;
       i.data.hasAction = i.hasAction || i.isCharged;
+      i.data.timelineLeftText = i.getTimelineTimeLeftDescriptive();
       i.data.showUnidentifiedData = i.showUnidentifiedData;
       if (i.showUnidentifiedData) i.data.name = getProperty(i.data, "data.unidentified.name") || getProperty(i.data, "data.identifiedName") || i.data.name;
       else i.data.name = getProperty(i.data, "data.identifiedName") || i.data.name;